requirements.txt
APScheduler==3.6.1
boto3==1.9.31
dj-database-url==0.5.0
django-storages==1.6.3
Django==3.1.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-crispy-forms==1.7.2
gunicorn==19.9.0
Pillow==5.3.0
psycopg2-binary==2.8.5
xmltodict==0.12.0
requests==2.20.0
